Setup using normalized homographies -  not successful since the flipping issue still exists: the localXYmap computed in videoplayer.cpp needs to be flipped  ..aber ich gehe jetzt erstmal pennen :)

+++ ARLib/include/ARLib/Webcam/VideoPlayer.h
@@ -28,7 +28,7 @@ private:
 	CCapture *capture;
 	float videoDistance;
 	ocam_model *ocamModel;
-	float homographyMatrix[3 * 3];
+	double homographyMatrix[3 * 3];
 };
 
 }; // ARLib namespace
